split task plugin and template interface
